feat: add scrim feature to matches/lan
