+++ app/Livewire/GameCustomHistoryTable.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Livewire;
 
+use App\Livewire\Traits\HasScrimEditor;
 use App\Models\Player;
 use Illuminate\View\View;
 use Livewire\Component;
@@ -11,30 +12,16 @@ use Livewire\WithPagination;
 
 class GameCustomHistoryTable extends Component
 {
+    use HasScrimEditor;
     use WithPagination;
 
     public Player $player;
-
-    public bool $isScrimEditor = false;
-
-    public array $scrimGameIds = [];
 
     // @phpstan-ignore-next-line
     public $listeners = [
         '$refresh',
         'toggleScrimMode',
     ];
-
-    public function toggleScrimMode(): void
-    {
-        $this->isScrimEditor = ! $this->isScrimEditor;
-    }
-
-    /** @codeCoverageIgnore */
-    public function updatedScrimGameIds(): void
-    {
-        $this->dispatch('syncGameIds', $this->scrimGameIds)->to(ScrimTogglePanel::class);
-    }
 
     public function paginationView(): string
     {
