+++ app/Livewire/GameHistoryTable.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Livewire;
 
+use App\Livewire\Traits\HasScrimEditor;
 use App\Models\Player;
 use Illuminate\View\View;
 use Livewire\Component;
@@ -11,6 +12,7 @@ use Livewire\WithPagination;
 
 class GameHistoryTable extends Component
 {
+    use HasScrimEditor;
     use WithPagination;
 
     public Player $player;
@@ -18,6 +20,7 @@ class GameHistoryTable extends Component
     // @phpstan-ignore-next-line
     public $listeners = [
         '$refresh',
+        'toggleScrimMode',
     ];
 
     public function paginationView(): string
@@ -28,6 +31,7 @@ class GameHistoryTable extends Component
     public function render(): View
     {
         return view('livewire.game-history-table', [
+            'isScrimEditor' => $this->isScrimEditor,
             'games' => $this->player
                 ->games()
                 ->with(['playlist', 'map', 'category'])
