+++ resources/views/pages/player.blade.php
@@ -12,7 +12,7 @@
                 <div class="notification">
                     <a class="is-small" href="{{ route('historyCsv', [$player, $type]) }}" rel="nofollow">export to csv</a>
 
-                    @if (in_array($type, ['custom']))
+                    @if (in_array($type, ['custom', 'matches', 'lan']))
                         <livewire:scrim-toggle-panel></livewire:scrim-toggle-panel>
                     @endif
                 </div>
